+++ Tournaments.Domain/Mapping/UserProfile.cs
@@ -1,10 +1,11 @@
 ﻿using AutoMapper;
 using Tournaments.Domain.Entities;
 using Tournaments.Domain.Models;
+using Tournaments.Domain.Models.AuthModels;
 
 namespace Tournaments.Domain.Mapping
 {
-	public class UserProfile : Profile
+    public class UserProfile : Profile
 	{
 		public UserProfile()
 		{
