+++ Tournaments.Domain/Models/AuthModels/AuthenticationResultModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Tournaments.Domain.Models.AuthModels
+{
+    public class AuthenticationResultModel
+    {
+        public string Token { get; set; } = null!;
+    }
+}
