+++ Tournaments.Domain/Models/AuthModels/LoginModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Tournaments.Domain.Models.AuthModels
+{
+    public class LoginModel
+    {
+        public string UserName { get; set; } = null!;
+        public string Password { get; set; } = null!;
+    }
+}
