+++ Tournaments.Domain/Models/AuthModels/RegisterModel.cs
@@ -1,0 +1,11 @@
+﻿namespace Tournaments.Domain.Models.AuthModels
+{
+    public class RegisterModel
+    {
+        public string UserName { get; set; } = null!;
+        public string Email { get; set; } = null!;
+        public string Password { get; set; } = null!;
+        public DateTime Birthday { get; set; }
+        public string? PhoneNumber { get; set; }
+    }
+}
