--- Tournaments.Domain/Models/AuthenticationResultModel.cs
@@ -1,7 +1,0 @@
-﻿namespace Tournaments.Domain.Models
-{
-    public class AuthenticationResultModel
-    {
-		public string Token { get; set; } = null!;
-	}
-}
