+++ Tournaments.Domain/Models/BracketModels/BracketModel.cs
@@ -1,0 +1,11 @@
+﻿using Tournaments.Domain.Entities;
+
+namespace Tournaments.Domain.Models.BracketModels
+{
+	public class BracketModel
+	{
+		public long Id { get; set; }
+		public long TournamentId { get; set; }
+		public List<Stage> Stages { get; set; } = null!;
+	}
+}
