--- Tournaments.Domain/Models/LoginModel.cs
@@ -1,8 +1,0 @@
-﻿namespace Tournaments.Domain.Models
-{
-	public class LoginModel
-	{
-        public string UserName { get; set; } = null!;
-		public string Password { get; set; } = null!;
-    }
-}
